backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter==23.1
PyJWT==2.1.0
requests==2.26.0
psycopg2-binary==2.9.3
djoser==2.1.0
Pillow==9.0.0
gunicorn==20.1.0
reportlab==4.0.4